refactor: 패키지 이동 및 사용하지 않는 클래스 삭제
