+++ tomcat/src/main/java/nextstep/jwp/service/UserService.java
@@ -1,4 +1,4 @@
-package nextstep.jwp;
+package nextstep.jwp.service;
 
 import java.util.NoSuchElementException;
 import nextstep.jwp.db.InMemoryUserRepository;
